Delete old courses  (#99)
* Init - SHOULD NOT PASS

* testing

* Paralell tests & code coverage

* Update path properly

* Tweaks

* Tweaks

* Fix classhash

* Should work now, enabling deletion

* trial without adding new

* verbose

* Lgging

* Fix tests

* Remove logging

* revert dependency

* Add back verbose

* Explicitly test sections

* logging

* termId

* logging

* Add crn

* Remove logs

* Address PR comments

* Revert flagged changes
