+++ tests/workflow_tests/setup.test.git.ts
@@ -11,7 +11,7 @@ const NUMS_COURSES = {
   "202260": 493,
 };
 const NUMS_SECTIONS = {
-  "202240": 839,
+  "202240": 840,
   "202250": 1634,
   "202260": 515,
 };
